+++ CHANGES.md
@@ -2,3 +2,6 @@
 
 ## 0.1
 - Initial release
+- added useState to store the title and description value
+- Added TextInput to show the text and edit the text instead of Text component
+- Added new library react-native-responsive-screen